requirements.txt
Flask==2.0.2
Flask_Bcrypt==0.7.1
flask_login==0.5.0
flask_migrate==3.1.0
flask_sqlalchemy==2.5.1
flask_wtf==1.0.0
SQLAlchemy==1.4.29
Werkzeug==2.0.2
WTForms==3.0.1
flask-markdown
email_validator